requirements.txt
flask==2.3.3
requests==2.31.0
openai==1.30.1
python-dotenv==1.0.1

langchain==0.1.16
langchain-community==0.0.35

sentence-transformers==2.7.0
faiss-cpu==1.8.0
SQLAlchemy==2.0.30